Create structure to add current players to the store.
Created additional files to add current players to the store, including a top level node connecting them.  This required moving some utility functions to a higher level where they could be more easily re-used.

Additionally, a pattern was sought to eliminate the need to re-implement selectors at every level of the state object.  Such a pattern was found and appears to be working, a wrapper function is used to trim the state object before it is passed to the root selectors in each file.  The potential pitfalls and limitations of this approach are not yet known.