Pin a working version of ajv.
The latest version of ajv, a schema parser that create-react-app relies on, has some incomparability with webpack that causes 'npm start' to fail.  The solution appears to be to pin a previous version, so that's what I've done.

This leads to some conflicts with wanted versions in NPM that will need to be monitored, but for now this is what allows work to move forward.