Added react router to project.
I had to delete some create-react-app default files, like the logo.svg, that were not working well when combined with react-router.  There don't seem to be any problems from removing them and it's not like the project particularly needed the default favorite icon.

I copied a simple login form example and I've been modifying it to look more like the Dojo's muster sign-in form.  I added react-form but it turns out that the process of configuring react-form is not trivial so I am checkpointing here.