Moving to Redux Duck style modules.  Simplifying selector pathing.
A single file containing reducers, selectors, action constants and action creators for the known players module has been created to reduce the number of small and widely dispersed files.

The path from the top level of the application's state tree is now specified in the selectors, but this complicates testing.
The approach currently used is to create a dummy reducer using CombineReducer to match the expected path.
This also means that the test data stores were updated to reflect this overall path.

I, David Perkins, am not happy with this approach but I can't think of any better alternatives at the moment.
I could allow the path provided to reducers to be modified at runtime but the easiest way to do this would be to have a non-constant module scoped variable for the path which I am trying to avoid.