Connected sign-in form, improved local storage
Submitting the sign in form now clears the form, adds a new player if required, and adds them to the set of current players.

Using the react-redux connect function to inject bound action creators is not working, so for the time being the dispatch prop injected into the submit function by redux-form is being used.

Because JSON  has limited support for data types like sets, it was necessary to update the local storage code to properly restore data from JSON to things like immutable Sets and Records.