Fixed ReduxForm and Redux.connect prop injection conflict
This was fixed by reordering the application of higher order components.  The order that works appears to be different from the order suggested in the ReduxForm documentation.

It isn't clear why the ReduxForm documentation picked one order over the other.  Redux.connect inject bound action creators and props derived from the state of the redux store into a component.  If you apply the Redux.connect HOC first, ReduxForm won't see those injected props.  That means they won't get passed to the props argument for all of ReduxForm's function callbacks, including onSubmit.

These functions do receive the store's dispatch function as an argument, so you can re-bind action creators inside them, but this leads to situations where it would be necessary to bind the same action creator twice for the same component, once with Redux.connect and again inside any ReduxForm callbacks that need to use it.

The worst possibility is that Redux.connect may cause a component to be redrawn, and ReduxForm might not, for some unknown reason, be able to handle that?  This would be visible if ReduxForm forms started losing their state for no other visible reason.  It is important to remember this risk.