Added the results from a spike integrating material-ui and React Beautiful DND.
These files are not appropriate for this project by themselves but will need to be used a model and transformed over time to provide the needed functionality.

Integrating the spike results into this project shows a regression.  It's not clear if that's because the latest version of the spike files were not used or because of something about the way this project is configured.  It will need to be investigated.