Recovering from the loss of a hard drive
It looks like I forgot to get push often enough, so some work was lost.
This creates a new client application and restores the fragments of work that did get saved.
It is on an isolated branch both because it is incomplete and because it isn't in a working state.

This also updates the project's dependencies, including a major version of socket.io
so there may be conflicts.