GitHub actions (#38)
Configure GitHub Actions to perform continuous integration tests on commits, pull requests to major branches

The workflow is configured to cache npm dependencies to speed runs and reduce network usage.

In order to facilitate this workflow, a _package.json_ file was created at the top level of the repository which includes a workspace dependency on the modules it contains.  This allows all tests to be run with `npm test` at the top level of the project, as the CI workflow expects.

Because there is now a top level package file, it has been configured to apply _prettier_ as a pre-commit hook using _Husky_.  This should help keep the project files formatted.